// MarkEd-front/src/api/models/PeerMatch.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PeerMatch = {
    reviewer_name: string;
    reviewer_userNumber: string;
    reviewer_email: string;
    submission_owner_name: string;
    submission_owner_userNumber: string;
    submission_owner_email: string;
    status: string;
};